refactor(src/): traits + hx-* modules
* ignore directories starting with underscore
* tweak file watch globs in `yarn start` script
* move `src/helix-ui/mixins` to `src/helix-ui/elements/traits`
* convert `HXPopoverElement.js` to `hx-popover` module
* convert `HXSearchAssistanceElement.js` to `hx-search-assistance` module
* convert `HXMenuElement.js` to `hx-menu` module
* convert `HXTooltipElement.js` to `hx-tooltip` module
