+++ .gitignore
@@ -11,6 +11,7 @@ public/
 .deploy*/
 dist/
 lib/
+_*/
 
 # Removing Coverage Reports folder created.
 coverage
