+++ src/helix-ui/elements/hx-tooltip/_shadow.less
@@ -1,10 +1,12 @@
-@import "./HXElement";
-@import "./_positioning";
+@import "../HXElement";
+@import "../_positioning";
 
-#hxPopover {
+#hxTooltip {
   background-color: var(--hxBackgroundColor, @gray-0);
   border-color: var(--hxBorderColor, @gray-300);
+  border-radius: 2px;
   border-style: solid;
   border-width: 1px;
   box-shadow: @layering-shadow;
+  padding: 0.75rem;
 }
