+++ src/helix-ui/elements/index.js
@@ -18,17 +18,17 @@ export { HXFileInputElement } from './HXFileInputElement';
 export { HXFileTileElement } from './HXFileTileElement';
 export { HXFormControlElement } from './HXFormControlElement';
 export { HXIconElement } from './HXIconElement';
-export { HXMenuElement } from './HXMenuElement';
+export { HXMenuElement } from './hx-menu/index.js';
 export { HXMenuitemElement } from './HXMenuitemElement';
 export { HXModalElement } from './HXModalElement';
 export { HXPillElement } from './HXPillElement';
-export { HXPopoverElement } from './HXPopoverElement';
+export { HXPopoverElement } from './hx-popover/index.js';
 export { HXProgressElement } from './HXProgressElement';
 export { HXRadioControlElement } from './HXRadioControlElement';
 export { HXRadioElement } from './HXRadioElement';
 export { HXRadioSetElement } from './HXRadioSetElement';
 export { HXRevealElement } from './HXRevealElement';
-export { HXSearchAssistanceElement } from './HXSearchAssistanceElement';
+export { HXSearchAssistanceElement } from './hx-search-assistance/index.js';
 export { HXSearchElement } from './HXSearchElement';
 export { HXSelectControlElement } from './HXSelectControlElement';
 export { HXSelectElement } from './HXSelectElement';
@@ -40,4 +40,4 @@ export { HXTabsetElement } from './HXTabsetElement';
 export { HXTextControlElement } from './HXTextControlElement';
 export { HXTextareaControlElement } from './HXTextareaControlElement';
 export { HXToastElement } from './HXToastElement';
-export { HXTooltipElement } from './HXTooltipElement';
+export { HXTooltipElement } from './hx-tooltip/index.js';
