+++ src/helix-ui/elements/traits/Revealable.js
